reorganize src directories and delete old icons
